First cut at CGI backend for ggcov
This will be the cornerstone of the new web2 architecture: a C++
CGI program which uses the cov code directly, instead of PHP scripts
which use a pre-built database.

Added a class json_t to encapsulate the serialising of a data structure
to JSON format.

Added a class cgi_t which wraps the Cgicc library, hiding the STLness.

Added a first primitive cut of cov_project_t which will be a new top-level
object for containing files etc.  Currently very hacky and hardcoded,
and there can be only one.

Implemented the "listfiles" query which lists the source files in a project,
and the "annotate" query which annotates a single source file.