Remove cov_file_t::unminimise_name() & its caller
this function is not ever needed, as we have a cov_file_t with
both minimal and maximal names at any time we want.  Rejig several
functions, notably the show_lines() methods of summarywin and
sourcewin, to take const cov_file_t * instead of a filename, thus
recognising and enforcing the reality.