Move cov_function_t::list_all() into cov_project_t.
And adjust its two callers.

This function is kinda dumb, it builds a huge linked list of almost
every function in the project, then sorts it; the result gets passed to
populate_function_combo() in two different places which does little
more than iterate over it.  A smarter arrangement would be to do
the gathering and sorting in a ptrarray, which could be done with a
single allocation.  But that's for next time.