Implemented improved Project Browser page
Per my paper design.  The list of projects is presented using a paged
format, with First, Prev, Next, and Last links which appear or not
according to which page we're on.  The whole interface is implemented
in JS in the browser, with the projects list being fetched with
an AJAX call the first time the page is shown.  Alternate rows are
coloured differently.  The name, description, and date are carefully
formatted to be more readable than the default.