Add the Lego Diagram to the Reports list
For the moment, this single diagram is appended by JS to the reports list
returned from the "listreports" query.  JS handles reports and diagrams
similarly in the list but shows separate windows on click.  For the diagram
window, we just have one big <img> and in JS we set src= to point to a new CGI
query "diagram" which builds a PNG file using gd_scenegen_t and emits it as the
reply body.