Split ggcov RPM packages into three.
All three are subpackages (ggcov-cli, ggcov-gnome and ggcov-web) and the
main package is no longer built.

Also, the configure options --disable-gui and --disable-web will disable
the corresponding subpackages in the ggcov.spec file, and build only
those parts of the source needed for the remaining subpackages.  In
particular this means we can build the ggcov-cli RPM package from source
on a server with no GNOME libraries installed.