Fix a really old bug with minimal_name()
It turns out that a recursive walk over a source tree with the wrong
combination of unbuilt files in the wrong places, could cause the
cov_file code to get confused and recalculate common_path_ prematurely,
resulting in an incorrect common_path_ and bizarre broken strings from
minimal_name().  This bug appears to have existed for the better part of
a decade.