Better classnames for Cobertura report
When building class names to include in the Cobertura report, don't just
report the filename.  The Jenkins Cobertura plugin really gets upset
with / in what it thinks is a Java classname.  So replace the / with .
and remove the .c suffix, resulting in something much more like a Java
class name.