Add ggcov-run.
Ggcov-run is a shell script wrapper which inserts a preload shim into a
test program.  The preload shim intercepts open() library calls and
redirects the open() used in the gcc --coverage instrumentation to write
.gcda files at program exit, such that the file is written to a new
location.  This is quite handy for those cases where the test program
does not have write access to the hardcoded .gcda pathnames.