Reimplement support for suppressing
Expand the use of cov_suppression_t so that any reason that a file,
function, block, arc, or line might be suppressed, can be expressed by a
single cov_suppression_t object.

Reimplemented all the hardcoded pathnames and function names which
trigger suppression as the builtin defaults for the suppressions table,
which are now initialised in one place.  The -X -Y and -Z commandline
options now just add entries to that table.

Replace the suppressed_ flag (and the similar common_ flag for files)
with a pointer to a cov_supression_t or NULL.  Remove the distinction
between self-suppressed or suppressed by other objects; an object is
either suppressed or not.  Do all suppression at file read time in the
finalise phase, so that is_suppressed() is always trivial later.

Each object gives itself a suppression at the earliest opportunity, for
example as soon as it receives a name which can be used to look up the
suppressions table.  Otherwise, it can be suppressed externally later,
e.g. lines are suppressed as a side effect of parsing C source for magic
comments.

Attempts to add a NULL suppression, or to add the 2nd or subsequent
non-NULL suppression are silently ignored.  Any time an object gains a
suppressions, it pushes that new suppression down to the object's
inferiors, e.g. suppressing a function imemdiately suppresses all the
function's blocks.  Objects also inherit their superior's suppression
when they are created, so e.g. when adding a new function to a
suppressed file the function will be automatically suppressed.

In the post-read finalise phase, suppressions are percolated upwards
using the rule that an object is suppressed if all of its inferiors is
suppressed.  Thus e.g. a file will be suppressed if all its functions
are suppressed.

This has several useful effects

 * the code is more orthogonal and easier to understand

 * adding new suppressions, even dynamically, will be easier

 * it's also significantly faster, due largely to the number of
   calls to cov_line_t::find() being halved, but also because the
   is_suppressed() method is always trivial now.

 * it is now trivial to discover *why* a particular object was
   suppressed (although the gui doesn't use that yet).