Check in agedata.pl & agemap.pl
I wrote these scripts about a year ago as prototypes for a method of
calculating and visualising a metric which approximates the bug
probability density of the various source files in a GIT repo.  I still
haven't implemented that visualisation in ggcov properly, but in the
meanwhile the perl scripts do work (albeit very slowly) and might be
useful to someone.