Add teststarter_t, a helper for end-to-end unit tests
This class can be used to conveniently build a directory full
of .c files and a runtest script, then run the script to get .gcno
and .gcda files, and then call cov_read_files().  Of course this
affects vast quantities of global library state, but the recent
change to running each test in a forked child avoids that issue.