Reimplement all ggcov argument parsing
using the new argparse module.  Class cov_project_param_t provides the
common commandline options and the list of files.  The dependency on the
libpopt library is still there but is hidden behind a C++ API (except
when we need to expose it to integrate with antique versions of the
GNOME libraries).  Each program provides its own subclass of
cov_project_param_t which extends it with the program's own options.
The result is not only more C++ clean but eliminates a number of clunky
global variables which were complicating regression testing of the cov
library as a whole.