Fix update of flow diagram in the Source window
to actually show the diagrams.  The update_flow() function was being
called too early, before the GtkTextView had calculated its geometry
fully, and was calculating a bogus zero pixe height for the flow
diagram, then caching the bogus result.  Fix is to delay the actual
work of update_flows() using an idle function.