Fix unit test failure which lost commandline option values
Sometime since this code was written, libpopt's behavior when invoking its
option callback for string-valued options changed from passing a pointer into
the original argv[] which is stable forever, to passing a pointer to some
internal buffer which contains a copy of the value from argv[] only just during
the life of the callback.  This was causing the simple argument parser to fail
unit tests.  Fixed by strdup()ing the value.